quantum-entanglement/requirements-dev.txt
qutip==5.0.1
qutip-qip==0.3.1
qutip-qtrl==0.1.1
sympy==1.12
mlflow==2.12.2
apache-airflow[celery]==2.9.1
apache-airflow-providers-docker
papermill
minio==7.2.7
#tensorflow==2.16.1